Rename Planner to Assistant
More in line with new design
