+++ src/screens/Assistant/ResultItem.tsx
@@ -2,11 +2,11 @@ import nb from 'date-fns/locale/nb';
 import React from 'react';
 import {Text, View, ViewStyle} from 'react-native';
 import Dash from 'react-native-dash';
-import WalkingPerson from '../../../assets/svg/WalkingPerson';
-import {Leg, LegMode, TripPattern} from '../../../sdk';
-import {StyleSheet} from '../../../theme';
-import colors from '../../../theme/colors';
-import {formatToClock, secondsToDuration} from '../../../utils/date';
+import WalkingPerson from '../../assets/svg/WalkingPerson';
+import {Leg, LegMode, TripPattern} from '../../sdk';
+import {StyleSheet} from '../../theme';
+import colors from '../../theme/colors';
+import {formatToClock, secondsToDuration} from '../../utils/date';
 import JourneyBusIcon from './svg/JourneyBusIcon';
 import JourneyTrainIcon from './svg/JourneyTrainIcon';
 import JourneyTramIcon from './svg/JourneyTramIcon';
