+++ src/screens/Assistant/svg/JourneyTrainIcon.tsx
