+++ src/screens/Assistant/svg/JourneyTramIcon.tsx
