+++ src/screens/Assistant/svg/LeftArrow.tsx
