+++ src/screens/Assistant/svg/RightArrow.tsx
